+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/multiplatform/_internal/configureMultiplatformWatchOS.kt
@@ -56,7 +56,7 @@ internal fun configureMultiplatformWatchOS(project: Project) {
             watchosTest.dependsOn(commonTest)
             if (darwinTest != null) watchosTest.dependsOn(darwinTest)
             for (watchosTestSourceSet in watchosTestSourceSets) {
-                watchosTestSourceSet.dependsOn(watchosMain)
+                watchosTestSourceSet.dependsOn(watchosTest)
             }
         }
     }