+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/extensions/kotlin/multiplatform/targets/KotlinMultiplatformiOSExtension.kt
@@ -11,7 +11,15 @@ public open class KotlinMultiplatformiOSExtension :
 
     override var Project.isEnabled: Boolean
         get() = hubdleState.kotlin.multiplatform.ios.isEnabled
-        set(value) = hubdleState.kotlin.multiplatform.ios.run { isEnabled = value }
+        set(value) =
+            hubdleState.kotlin.multiplatform.run {
+                darwin.isEnabled = value
+                ios.isEnabled = value
+                iosArm32.isEnabled = value
+                iosArm64.isEnabled = value
+                iosSimulatorArm64.isEnabled = value
+                iosX64.isEnabled = value
+            }
 
     override val name: String = "ios"
 }